Adaptation of rust for openEuler
maillist inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I5J75G
CVE: NA

--------------------------------

The version of openeuler is different from the latest
version of the mainline, so need adapted

+++ rust/kernel/random.rs
@@ -17,7 +17,7 @@ pub fn getrandom(dest: &mut [u8]) -> Result {
     }
 
     unsafe {
-        bindings::get_random_bytes(dest.as_mut_ptr() as *mut c_types::c_void, dest.len());
+        bindings::get_random_bytes(dest.as_mut_ptr() as *mut c_types::c_void, dest.len() as i32);
     }
     Ok(())
 }
@@ -37,6 +37,6 @@ pub fn getrandom_nonblock(dest: &mut [u8]) -> Result {
 /// Does *not* credit the kernel entropy counter though.
 pub fn add_randomness(data: &[u8]) {
     unsafe {
-        bindings::add_device_randomness(data.as_ptr() as *const c_types::c_void, data.len());
+        bindings::add_device_randomness(data.as_ptr() as *const c_types::c_void, data.len() as u32);
     }
 }
